add adjustments shader
added a fragment shader that can do multiple basic adjustments:
gamma,
desaturation,
inversion,
contrast, and
brightness.

They are applied in that order so earlier adjustments affect later ones.

This shader can be used with or without a texture.
With a texture, the colour is multiplied with image as normal but is applied directly after desaturation, allowing colorization by desaturation and the multiplying the colour.
Without a texture, the object's colour is its initial value instead of the texture value. This also means that the "colorize step" is skipped since the colour is already used for the main colour.

Each adjustment can have its own value that specifies the amount of effect.
Desaturaion and inversion both range from 0 (no effect) to 1 (full effect).
Contrast and brightness both range from -1 (full effect in negative direction) through 0 (no effect) to 1 (full effect in positive direction).
Gamma has not upper limit although, practically, there will be very little useful or noticable effect difference after 100 or so. Its lower limit is that it must be greater than 0. A value of 1 is 'no change'. Gamma is provided mainly as a 'gamma correction' tool, which is why it is first in the chain - to allow textures to corrected beforehand.

The texture should be sent to the uniform "texture" and the uniform "useTexture" is a boolean that controls whether or not to use the texture.